move server definition to package protocols
